- Adding units namespace with units conversion factors.  - Adding CMolecule class storing and manipulating single molecule geometry and properties.  - Adding CMolXYZReader class for reading @molxyz control group.

+++ src/readers/EnvironmentReader.cpp
@@ -39,7 +39,7 @@ void CEnvironmentReader::parse(const CInputData&    inputData,
 
         auto ncommands = cgroup.getNumberOfCommands();
 
-        for (size_t i = 0; i < ncommands; i++)
+        for (int32_t i = 0; i < ncommands; i++)
         {
             auto iline = cgroup.getCommand(i);
