+++ src/readers/JobsReader.cpp
@@ -40,7 +40,7 @@ void CJobsReader::parse(std::vector<int32_t>& listOfJobIds,
 
         auto ncommands = cgroup.getNumberOfCommands();
 
-        for (size_t i = 0; i < ncommands; i++)
+        for (int32_t i = 0; i < ncommands; i++)
         {
             auto iline = cgroup.getCommand(i);
